ZSET B+ Tree PR 1: extract skiplist into dedicated module (#3579)
This is PR 1 of the series introducing an abstraction layer over the
sorted set skiplist, enabling a future swap to a B+ tree for memory and
performance gains. (Issue #3166) It targets the zset-btree feature
branch.

### Planned PRs:
1. **Skiplist extraction ← this PR**
2. OrderedIndex interface + skiplist backend + unit tests
3. Convert t_zset.c to OrderedIndex
4. Convert module.c
5. Convert remaining files
6. Integration tests

### Summary of changes:

Move the skiplist implementation out of `t_zset.c` into its own
compilation unit (`skiplist.c` / `skiplist.h`), establishing a clean
module boundary.

**New files:**
- `src/skiplist.h` — struct definitions (`zskiplistNode`, `zskiplist`),
inline helpers (span/height accessors), and function declarations
- `src/skiplist.c` — all skiplist data structure operations extracted
from `t_zset.c`

**Changes to `server.h`:**
- Remove `zskiplistNode` and `zskiplist` struct definitions (moved to
`skiplist.h`)
- Remove `ZSKIPLIST_MAXLEVEL`, `ZSKIPLIST_MAX_SEARCH` defines (moved to
`skiplist.h`)
- Remove skiplist function declarations (moved to `skiplist.h`)
- Add forward declaration `struct zskiplist` for the `zset` struct
pointer

**Changes to `t_zset.c`:**
- Remove skiplist implementation (now in `skiplist.c`)
- Add `#include "skiplist.h"`
- Retain range parsing helpers (`zslParseRange`, `zslParseLexRangeItem`,
`zsetParseLexRange`, `zsetFreeLexRange`) as command-layer logic

**Other files:**
- Add `#include "skiplist.h"` to files that dereference skiplist struct
fields (`defrag.c`, `geo.c`, `module.c`, `object.c`, `sort.c`, `aof.c`,
`db.c`, `debug.c`, `rdb.c`, `server.c`, `lazyfree.c`,
`valkey-check-rdb.c`)
- Update `Makefile` and `CMakeLists.txt` to compile `skiplist.c`

Functions that were `static` in `t_zset.c` and are now needed across
translation units are made non-static. No behavioral changes.
