+++ src/Makefile
@@ -567,6 +567,7 @@ ENGINE_SERVER_OBJ = \
     t_stream.o \
     t_string.o \
     t_zset.o \
+    skiplist.o \
     threads_mngr.o \
     timeout.o \
     tls.o \
