+++ src/debug.c
@@ -28,6 +28,7 @@
  */
 
 #include "server.h"
+#include "skiplist.h"
 #include "util.h"
 #include "sha1.h" /* SHA1 is used for DEBUG DIGEST */
 #include "crc64.h"
